Moved exercise-app-state to ExerciseManager
